Deleting the unused class org.sonar.plugins.scala.Scala.scala and refactoring the class sagacify.scala.ScalaPlugin
